+++ changes/CA-2746.other
@@ -1,0 +1,1 @@
+Split upgrade with high memory consumption into two. [deiferni]
